+++ src/dist_s1/__main__.py
@@ -7,14 +7,13 @@ from .data_models.runconfig_model import RunConfigData
 from .workflows import run_dist_s1_sas_workflow, run_dist_s1_workflow
 
 
-def localize_data(mgrs_tile_id: str, post_date: str | datetime, track: int, post_buffer_days: int):
-    """Dummy function to localize data."""
+def localize_data(mgrs_tile_id: str, post_date: str | datetime, track: int, post_buffer_days: int) -> dict:
     click.echo('Localizing data')
     return {'mgrs_tile_id': mgrs_tile_id, 'post_date': post_date, 'track': track, 'buffer_days': post_buffer_days}
 
 
 @click.group()
-def cli():
+def cli() -> None:
     """CLI for dist-s1 workflows."""
     pass
 
@@ -22,7 +21,7 @@ def cli():
 # SAS Workflow (No Internet Access)
 @cli.command(name='run_sas')
 @click.option('--runconfig_yml_path', required=True, help='Path to YAML runconfig file', type=click.Path(exists=True))
-def run_sas(runconfig_yml_path: str | Path):
+def run_sas(runconfig_yml_path: str | Path) -> str:
     runconfig_data = RunConfigData.from_yaml(runconfig_yml_path)
     out_dir_data = run_dist_s1_sas_workflow(runconfig_data)
     click.echo(f'Writing to DIST-S1 product to directory: {out_dir_data.path}')
@@ -42,7 +41,7 @@ def run_sas(runconfig_yml_path: str | Path):
     'Near the dateline you may have two sequential track numbers.',
 )
 @click.option('--post_date_buffer_days', type=int, required=True, help='Buffer days around post-date.')
-def run(mgrs_tile_id: str, post_date: str, track_number: int, post_date_buffer_days: int):
+def run(mgrs_tile_id: str, post_date: str, track_number: int, post_date_buffer_days: int) -> str:
     """Localize data and run dist_s1_workflow."""
     # Localize data
     run_config = run_dist_s1_workflow(mgrs_tile_id, post_date, track_number, post_date_buffer_days)